Prevent email spamming via sending a request to `/auth/generate` every 60 seconds (#7)
* :lock: prevent emails from email spamming with otp

Modified the schema to contain a attribute that counts the number of times an account has been sent an email. With this, the cooldown is doubled each time that email address receives an email. The exponential growth of a cooldown period should effectively reduce the issue of spamming.

* :bug: small fix to eliminate potential race condition

Use `$inc` MongoDB operator to decrement remainingAttempts