Event Pass Integration (#31)
* Event Pass Integration

This commit adds multiple changes, including:

• Migrating the parameter setup for the Event Pass function from the Wallet Controller to the Wallet Service
• Added a call to Wallet Service in POST /attendee, including the Pass URL in the response of the endpoint
• Added a call to Wallet Service in the /auth/me endpoint, including the Pass URL in the response of the endpoint
• Added general formatting

* Fixed Build Issues

Ran `yarn test` to resolve import issues in wallet.controller and wallet.service as well as remove extraneous imports

* Updated README

Added the note that the Google Wallet Variables should be in double-quotes in the README documentation for the .env variables

* Updated Service for new Wallet Account

Changed high-level event class details in the Service file for our new Google Wallet Pass Service Account

* Updated Wallet Assets

Final Asset Tweaks to the Google Wallet Pass class to enhance the event pass aesthetics. Additionally, added the attendee email as a field on the pass itself for manual email validation.